fix: Batch Barton round-2 fixes (#961, #960, #955)
- #961: Extract combat narration string arrays to Data/CombatNarration.cs
  and merge with Systems/CombatNarration.cs ambient flavor text
- #960: Use Random.Shared instead of new Random() in CombatEngine;
  add optional Random? rng parameter to DungeonGenerator
- #955: Add XML doc comments to IDisplayService documenting
  input-coupled methods for future HUD/Dashboard refactor
